Another attempt at updating Podspec to make gigya into a true cocoapod depending upon other cocoapods
The problem is that the frameworks that are included here are point-releases, and not readily available as matched cocoapods.

When I try to specify dependencies within the gigya-sdk.podspec, then cocoapods complains about mismatched frameworks.

When I try to specify dependencies within the including project, then gigya still can not find LineSDK and GoogleSignIn.

I am pausing development here for now…